function-app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-search-documents==11.4.0b8
azure-identity
azure-functions
openai
langchain
tiktoken